master/requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
boto3==1.12.41
botocore==1.15.41
celery==4.4.2
Django==3.0.5
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
kombu==4.6.8
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
